# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/hollroom/monthly/domain/dto/MonthlyProductResponseDTO.java
@@ -21,7 +21,7 @@ public class MonthlyProductResponseDTO {
     private int monthly;
     private Long divisionCode;
     private int floorCount;
-    private int pyeongCount;
+    private Float pyeongCount;
     private int roomCount;
     private int bayCount;
     private String roomOption;
